requirements.txt
Django==1.11.16
django-cors-headers==2.4.0
django-rest-auth==0.9.3
djangorestframework==3.8.2
pytz==2018.5
six==1.11.0
